update: `my-investments` folder struct and component renamed to `investments`
