+++ src/app/infra/gateways/investments/investments-gateway.model.ts
