+++ src/app/infra/gateways/investments/investments-gateway.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { InvestmentsGatewayService } from './investments-gateway.service';
+
+describe('MyInvestmentsGatewayService', () => {
+	let service: InvestmentsGatewayService;
+
+	beforeEach(() => {
+		TestBed.configureTestingModule({});
+		service = TestBed.inject(InvestmentsGatewayService);
+	});
+
+	it('should be created', () => {
+		expect(service).toBeTruthy();
+	});
+});
