--- src/app/infra/gateways/my-investments/my-investments-gateway.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { MyInvestmentsGatewayService } from './my-investments-gateway.service';
-
-describe('MyInvestmentsGatewayService', () => {
-  let service: MyInvestmentsGatewayService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(MyInvestmentsGatewayService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
