+++ src/app/pages/restricted/investments/investments-dash/investments-dash.component.html
@@ -1,11 +1,11 @@
 <main class="min-h-screen bg-zinc-50 dark:bg-slate-900">
 	<section class="z-2 flex min-h-screen flex-row items-stretch gap-0 px-10 pt-20 desktop-2k:container">
-		<div
-			class="relative my-8 flex max-h-[calc(100vh-144px)] min-h-[600px] w-[350px] select-none flex-col gap-4 overflow-y-auto border-r border-zinc-300/40 p-4 dark:border-slate-50/[0.06]"
+		<section
+			class="relative my-8 flex max-h-[calc(100vh-144px)] min-h-[600px] w-[350px] flex-col gap-4 overflow-y-auto border-r border-zinc-300/40 p-4 dark:border-slate-50/[0.06]"
 		>
 			<button
 				type="button"
-				class="group/link flex flex-row items-center justify-center rounded-lg bg-zinc-100 px-3 py-2 text-sm transition-transform hover:-translate-y-0.5 active:scale-95 dark:bg-slate-900"
+				class="group/link flex select-none flex-row items-center justify-center rounded-lg bg-zinc-100 px-3 py-2 text-sm transition-transform hover:-translate-y-0.5 active:scale-95 dark:bg-slate-900"
 				routerLink="/new"
 				[disabled]="this.loadingWallets()"
 			>
@@ -25,7 +25,7 @@
 			} @else if (this.wallets().length) {
 				@for (wallet of this.wallets(); track wallet.wallet_id) {
 					<div
-						class="cursor-pointer rounded-lg border border-zinc-300/40 p-4 dark:border-slate-50/[0.06]"
+						class="cursor-pointer select-none rounded-lg border border-zinc-300/40 p-4 dark:border-slate-50/[0.06]"
 						[class]="{
 							'ring-2': this.selected_wallets().has(wallet.wallet_id),
 							'ring-dongs-400/65': this.selected_wallets().has(wallet.wallet_id),
@@ -115,13 +115,13 @@
 					<div class="flex select-none flex-col items-center justify-center p-6 text-xl font-semibold text-slate-400">No Wallets</div>
 				</div>
 			}
-		</div>
-		<div class="my-8 flex flex-1 select-none flex-col p-4">
-			<div class="flex flex-row items-center justify-between">
+		</section>
+		<section class="my-8 flex flex-1 flex-col p-4">
+			<menu class="flex flex-row items-center justify-between">
 				<div class="flex flex-row gap-2">
 					<button
 						type="button"
-						class="group/link flex flex-row items-center justify-center rounded-lg bg-zinc-100 px-3 py-2 text-sm transition-transform hover:-translate-y-0.5 active:scale-95 dark:bg-slate-900"
+						class="group/link flex select-none flex-row items-center justify-center rounded-lg bg-zinc-100 px-3 py-2 text-sm transition-transform hover:-translate-y-0.5 active:scale-95 dark:bg-slate-900"
 						[class]="{
 							'ring-2': this.selected_panel() === 'performance',
 							'ring-dongs-500/65': this.selected_panel() === 'performance',
@@ -136,7 +136,7 @@
 					</button>
 					<button
 						type="button"
-						class="group/link flex flex-row items-center justify-center rounded-lg bg-zinc-100 px-3 py-2 text-sm transition-transform hover:-translate-y-0.5 active:scale-95 dark:bg-slate-900"
+						class="group/link flex select-none flex-row items-center justify-center rounded-lg bg-zinc-100 px-3 py-2 text-sm transition-transform hover:-translate-y-0.5 active:scale-95 dark:bg-slate-900"
 						[class]="{
 							'ring-2': this.selected_panel() === 'balance_history',
 							'ring-dongs-500/65': this.selected_panel() === 'balance_history',
@@ -151,7 +151,7 @@
 					</button>
 					<button
 						type="button"
-						class="group/link flex flex-row items-center justify-center rounded-lg bg-zinc-100 px-3 py-2 text-sm transition-transform hover:-translate-y-0.5 active:scale-95 dark:bg-slate-900"
+						class="group/link flex select-none flex-row items-center justify-center rounded-lg bg-zinc-100 px-3 py-2 text-sm transition-transform hover:-translate-y-0.5 active:scale-95 dark:bg-slate-900"
 						[class]="{
 							'ring-2': this.selected_panel() === 'assets',
 							'ring-dongs-500/65': this.selected_panel() === 'assets',
@@ -166,23 +166,24 @@
 					</button>
 				</div>
 				<div class="flex flex-row gap-2">
-                    <div class="flex flex-row">
-                        <div class="flex flex-row items-center border-r border-dashed rounded-l-lg bg-zinc-100 px-3 py-2 text-xs dark:bg-slate-900">
-                            <span class="font-medium text-dongs-500 capitalize">{{ selected_currency() | lowercase }}</span>
-                        </div>
-                        <button
-                            type="button"
-                            class="group/link flex flex-row items-center justify-center rounded-r-lg bg-zinc-100 px-3 py-2 text-sm transition-transform active:scale-95 dark:bg-slate-900"
-                        >
-                            <fa-icon
-                                class="mr-2 group-hover/link:text-dongs-500"
-                                [icon]="icons().faCaretDown"
-                            ></fa-icon>
-                            <span class="font-medium group-hover/link:text-dongs-500">Currency</span>
-                        </button>
-                    </div>
+					<div class="flex flex-row">
+						<div class="flex flex-row items-center rounded-l-lg border-r border-dashed bg-zinc-100 px-3 py-2 text-xs dark:bg-slate-900">
+							<span class="font-medium capitalize text-dongs-500">{{ selected_currency() | lowercase }}</span>
+						</div>
+						<button
+							type="button"
+							class="group/link flex select-none flex-row items-center justify-center rounded-r-lg bg-zinc-100 px-3 py-2 text-sm transition-transform active:scale-95 dark:bg-slate-900"
+						>
+							<fa-icon
+								class="mr-2 group-hover/link:text-dongs-500"
+								[icon]="icons().faCaretDown"
+							></fa-icon>
+							<span class="font-medium group-hover/link:text-dongs-500">Currency</span>
+						</button>
+					</div>
 				</div>
-			</div>
-		</div>
+			</menu>
+			<section class="mt-4 flex flex-col gap-2"></section>
+		</section>
 	</section>
 </main>
