+++ src/app/pages/restricted/investments/investments.routes.ts
@@ -2,7 +2,7 @@ import { Routes } from '@angular/router';
 import { titleResolver } from '../../../app.routes';
 import { permissionGuard } from '../../../infra/guards/permission.guard';
 
-export const myInvestmentsRoutes: Routes = [
+export const investmentsRoutes: Routes = [
 	{
 		path: '',
 		data: {
@@ -11,7 +11,7 @@ export const myInvestmentsRoutes: Routes = [
 			permission: 'INVESTMENTS_ACCESS',
 		},
 		title: titleResolver,
-		loadComponent: () => import('./my-investments-dash/my-investments-dash.component').then((module) => module.MyInvestmentsDashComponent),
+		loadComponent: () => import('./investments-dash/investments-dash.component').then((module) => module.InvestmentsDashComponent),
 		canMatch: [permissionGuard],
 	},
 	{ path: '**', redirectTo: '', pathMatch: 'full' },
